2011-03-24 Roman - !!!!:SYS, DAQ.JavaLikeCalc: The User's system API regular expression object TRegExp, based on PCRE, support is added.     Realised functions RegExp.exec() and RegExp.test() standard JavaScript interface.     To DAQ.JavaLikeCalc is realised "new RegExp()" for TRegExp object create.     To DAQ.JavaLikeCalc addon string functions with using RegExp is realised: 	int search(); Array match(); Array split(RegExp pat, int limit); string replace(string substr, string str); 	string replace(RegExp pat, string str);. - !!:SYS: PCRECPP dependence is replaced by PCRE. - SYS: GD library checking for allow is moved to once. - SYS: The User's system API, TArrayObj is expanded for object's properties, not numbers, support.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@1313 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tsys.cpp
@@ -790,7 +790,7 @@ string TSYS::strEncode( const string &in, TSYS::Code tp, const string &symb )
 	    {
 		if(i_sz && !(i_sz%57))	sout.push_back('\n');
 		sout.push_back(base64alph[(unsigned char)in[i_sz]>>2]);
-		if((i_sz+1) >= in.size())
+		if((i_sz+1) >= (int)in.size())
 		{
 		    sout.push_back(base64alph[((unsigned char)in[i_sz]&0x03)<<4]);
 		    sout+="==";
@@ -798,7 +798,7 @@ string TSYS::strEncode( const string &in, TSYS::Code tp, const string &symb )
 		else
 		{
 		    sout.push_back(base64alph[(((unsigned char)in[i_sz]&0x03)<<4)|((unsigned char)in[i_sz+1]>>4)]);
-		    if((i_sz+2) >= in.size())
+		    if((i_sz+2) >= (int)in.size())
 		    {
 			sout.push_back(base64alph[((unsigned char)in[i_sz+1]&0x0F)<<2]);
 			sout.push_back('=');
@@ -847,7 +847,7 @@ string TSYS::strEncode( const string &in, TSYS::Code tp, const string &symb )
 	case TSYS::ShieldSimb:
 	    sout.reserve(in.size());
 	    for( i_sz = 0; i_sz < (int)in.size(); i_sz++ )
-		if( in[i_sz] == '\\' && i_sz < (in.size()-1) )
+		if( in[i_sz] == '\\' && i_sz < ((int)in.size()-1) )
 		{
 		    switch( in[i_sz+1] )
 		    {
